refactor: Add checkAdmin endpoint to RoomController
This commit adds a new endpoint to the RoomController called checkAdmin. This endpoint allows users to check if they are an admin of a specific room. The checkAdmin endpoint calls the checkRoomAdmin method in the RoomService to determine if the user is an admin of the room. The result is then returned as a JSON object indicating whether the user is an admin or not.

Ref: #123